refactor: mock error handler service
