+++ src/app/shared/components/template/services/template-field.service.spec.ts
@@ -4,7 +4,7 @@ import { TemplateFieldService } from "./template-field.service";
 import type { PromiseExtended } from "dexie";
 import { booleanStringToBoolean } from "src/app/shared/utils";
 import { ErrorHandlerService } from "src/app/shared/services/error-handler/error-handler.service";
-import { MockErrorHandlerService } from "src/app/shared/services/error-handler/error-handler.service.spec";
+import { MockErrorHandlerService } from "src/app/shared/services/error-handler/error-handler.service.mock.spec";
 
 /** Mock calls for field values from the template field service to return test data */
 export class MockTemplateFieldService implements Partial<TemplateFieldService> {
