+++ src/app/shared/services/error-handler/error-handler.service.spec.ts
@@ -3,16 +3,6 @@ import { TestBed } from "@angular/core/testing";
 import { ErrorHandlerService } from "./error-handler.service";
 import { FirebaseService } from "../firebase/firebase.service";
 
-/** Mock calls for sheets from the appData service to return test data */
-export class MockErrorHandlerService implements Partial<ErrorHandlerService> {
-  public logError(error: Error): Promise<void> {
-    throw error;
-  }
-  public handleError(error: Error): Promise<void> {
-    throw error;
-  }
-}
-
 describe("ErrorHandlerService", () => {
   let service: ErrorHandlerService;
